feat: Make ActionBar float when scrolled
ActionBar now upon mounting will check the height of the top banner.
If user is scrolled down past that point, the ActionBar will gain class
'is-loading', and that will make it position fixed. This gives the user constant
access to the controls within it.

Also made some CSS tweaks for better displaying on 13in laptop.