feat: printable class Roster
At request of registrar, added feature to print out a roster of a class
that lists all the students and their picture in a format friendly to the
professor learning names and faces.
Also refactored to use a new FullPageLoading component in the
printables, and not use  CSS to show/hide things.