refactor: Switch to /enrollments Canvas endpoint
The /students endpoint did not provide any data about the nature of a user's
enrollment in the course, and after talking with the registrar, this will not
be sufficient. Moved to the /enrollments URL instead. Code has been
refactored to allow for extra information about the enrollment relationship.

beyond simply the seat number.