fix: Data type wrong in is_in_AIS
In Student resource file, is_in_AIS comes back as integer, as it should,
on local server, but on Dev, it was coming back as string. Casting it
as an int now before it ever comes through the api.

NOTE! Also did the same fix to is_namespot_addition. Shouldn't
change anything, but if that becomes weird, keep this change in mind.