Overwite all student photos every photo fetch call
AIS does update their photo DB from time to time. We saw placeholder
images earlier, and many of those are now real student pictures.
Because of this, we need FetchPhotoRoster to download the photo data
to the server every time for every photo. Previously it checked "do we have
a photo for this student already?" and did not proceed if we did. So we
eliminate that check and re-save the photo with the new data every time.

Also adding in an extra 'test student' filter in Canvas fetch cause one
is spelled like that.