fix: Variable type on AIS enrollment results
The check to see if UC_CLASS_ROSTER_TBL asks if it's an array, and then
the elseif should be on it as well, not just on body.
Just body will always return true, and it's expected the UC_CLASS_ROSTER_TBL
will always be an object if it's not an array, so this should't change anything
but it's more correct this way.